Pointers over channels. Default map init. Deadlock removed.
The previous code had a bug where we attempt to send the inference
result down the MQTT channel even when we don't use MQTT at all. This
correctly causes the program to block. We are now checking if the
MQTT channel has been initialized which is only the case if we use MQTT.

We now also send *pointers* over channels rather than copies of values
which makes the program less memory extensive and thus less prone to
running out of memory.

Finally, we now correctly initialize Result Detections map before
attempting to access its keys when String() is called before the map was
even initialized leading to program panicing.