Big refactoring of the codee. More readable and testable.
Large parts of the code have been refactored to make them more readable
and arguably more testable.

A lot of the main() function boilerplate has been moved to separate
functions so we can easily write unit tests for them.

Some of the nomenclature has been changed:
We now prefer to call data collector as publisher making it more generic
in case in the future someone will want to implement a generic Publisher
interface that will be publishing to any data source rather than being
tied to MQTT

We now explicitly require publish flag to signal the intent that we want
to publish the data to remote analytics server.